Use double in RayTracer

+++ src/main/java/com/mikedeejay2/mikedeejay2lib/util/raytrace/RayTracer.java
@@ -44,7 +44,7 @@ public final class RayTracer
      * @param maxDistance The maximum distance that should be raytraced before it stops trying
      * @return The Vector of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Vector rayTraceVector(LivingEntity entity, int maxDistance)
+    public static Vector rayTraceVector(LivingEntity entity, double maxDistance)
     {
         return rayTraceVector(entity, maxDistance, FluidCollisionMode.NEVER, false);
     }
@@ -60,7 +60,7 @@ public final class RayTracer
      * @param ignorePassableBlocks If the raytrace should ignore blocks that are passable (grass, flowers, etc)
      * @return The Vector of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Vector rayTraceVector(LivingEntity entity, int maxDistance, boolean ignorePassableBlocks)
+    public static Vector rayTraceVector(LivingEntity entity, double maxDistance, boolean ignorePassableBlocks)
     {
         return rayTraceVector(entity, maxDistance, FluidCollisionMode.NEVER, ignorePassableBlocks);
     }
@@ -76,7 +76,7 @@ public final class RayTracer
      * @param fluidCollisionMode The fluid collision mode to use
      * @return The Vector of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Vector rayTraceVector(LivingEntity entity, int maxDistance, FluidCollisionMode fluidCollisionMode)
+    public static Vector rayTraceVector(LivingEntity entity, double maxDistance, FluidCollisionMode fluidCollisionMode)
     {
         return rayTraceVector(entity, maxDistance, fluidCollisionMode, false);
     }
@@ -91,7 +91,7 @@ public final class RayTracer
      * @param ignorePassableBlocks If the raytrace should ignore blocks that are passable (grass, flowers, etc)
      * @return The Vector of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Vector rayTraceVector(LivingEntity entity, int maxDistance, FluidCollisionMode fluidCollisionMode, boolean ignorePassableBlocks)
+    public static Vector rayTraceVector(LivingEntity entity, double maxDistance, FluidCollisionMode fluidCollisionMode, boolean ignorePassableBlocks)
     {
         World world = entity.getWorld();
         Location startLoc = entity.getEyeLocation();
@@ -129,7 +129,7 @@ public final class RayTracer
      * @param maxDistance The maximum distance that should be raytraced before it stops trying
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Location rayTraceLocation(LivingEntity entity, int maxDistance)
+    public static Location rayTraceLocation(LivingEntity entity, double maxDistance)
     {
         return rayTraceLocation(entity, maxDistance, FluidCollisionMode.NEVER, false);
     }
@@ -145,7 +145,7 @@ public final class RayTracer
      * @param ignorePassableBlocks If the raytrace should ignore blocks that are passable (grass, flowers, etc)
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Location rayTraceLocation(LivingEntity entity, int maxDistance, boolean ignorePassableBlocks)
+    public static Location rayTraceLocation(LivingEntity entity, double maxDistance, boolean ignorePassableBlocks)
     {
         return rayTraceLocation(entity, maxDistance, FluidCollisionMode.NEVER, ignorePassableBlocks);
     }
@@ -161,7 +161,7 @@ public final class RayTracer
      * @param fluidCollisionMode The fluid collision mode to use
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Location rayTraceLocation(LivingEntity entity, int maxDistance, FluidCollisionMode fluidCollisionMode)
+    public static Location rayTraceLocation(LivingEntity entity, double maxDistance, FluidCollisionMode fluidCollisionMode)
     {
         return rayTraceLocation(entity, maxDistance, fluidCollisionMode, false);
     }
@@ -176,7 +176,7 @@ public final class RayTracer
      * @param ignorePassableBlocks If the raytrace should ignore blocks that are passable (grass, flowers, etc)
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Location rayTraceLocation(LivingEntity entity, int maxDistance, FluidCollisionMode fluidCollisionMode, boolean ignorePassableBlocks)
+    public static Location rayTraceLocation(LivingEntity entity, double maxDistance, FluidCollisionMode fluidCollisionMode, boolean ignorePassableBlocks)
     {
         World world = entity.getWorld();
         Location entityLoc = entity.getEyeLocation();
@@ -215,7 +215,7 @@ public final class RayTracer
      * @param maxDistance The maximum distance that should be raytraced before it stops trying
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Vector rayTraceVector(Location location, int maxDistance)
+    public static Vector rayTraceVector(Location location, double maxDistance)
     {
         return rayTraceVector(location, maxDistance, FluidCollisionMode.NEVER, false);
     }
@@ -231,7 +231,7 @@ public final class RayTracer
      * @param fluidCollisionMode The fluid collision mode to use
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Vector rayTraceVector(Location location, int maxDistance, FluidCollisionMode fluidCollisionMode)
+    public static Vector rayTraceVector(Location location, double maxDistance, FluidCollisionMode fluidCollisionMode)
     {
         return rayTraceVector(location, maxDistance, fluidCollisionMode, false);
     }
@@ -247,7 +247,7 @@ public final class RayTracer
      * @param ignorePassableBlocks If the raytrace should ignore blocks that are passable (grass, flowers, etc)
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Vector rayTraceVector(Location location, int maxDistance, boolean ignorePassableBlocks)
+    public static Vector rayTraceVector(Location location, double maxDistance, boolean ignorePassableBlocks)
     {
         return rayTraceVector(location, maxDistance, FluidCollisionMode.NEVER, ignorePassableBlocks);
     }
@@ -262,7 +262,7 @@ public final class RayTracer
      * @param ignorePassableBlocks If the raytrace should ignore blocks that are passable (grass, flowers, etc)
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Vector rayTraceVector(Location location, int maxDistance, FluidCollisionMode fluidCollisionMode, boolean ignorePassableBlocks)
+    public static Vector rayTraceVector(Location location, double maxDistance, FluidCollisionMode fluidCollisionMode, boolean ignorePassableBlocks)
     {
         World world = location.getWorld();
         Vector lookVec = location.getDirection();
@@ -298,7 +298,7 @@ public final class RayTracer
      * @param maxDistance The maximum distance that should be raytraced before it stops trying
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Location rayTraceLocation(Location location, int maxDistance)
+    public static Location rayTraceLocation(Location location, double maxDistance)
     {
         return rayTraceLocation(location, maxDistance, FluidCollisionMode.NEVER, false);
     }
@@ -314,7 +314,7 @@ public final class RayTracer
      * @param fluidCollisionMode The fluid collision mode to use
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Location rayTraceLocation(Location location, int maxDistance, FluidCollisionMode fluidCollisionMode)
+    public static Location rayTraceLocation(Location location, double maxDistance, FluidCollisionMode fluidCollisionMode)
     {
         return rayTraceLocation(location, maxDistance, fluidCollisionMode, false);
     }
@@ -330,7 +330,7 @@ public final class RayTracer
      * @param ignorePassableBlocks If the raytrace should ignore blocks that are passable (grass, flowers, etc)
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Location rayTraceLocation(Location location, int maxDistance, boolean ignorePassableBlocks)
+    public static Location rayTraceLocation(Location location, double maxDistance, boolean ignorePassableBlocks)
     {
         return rayTraceLocation(location, maxDistance, FluidCollisionMode.NEVER, ignorePassableBlocks);
     }
@@ -345,7 +345,7 @@ public final class RayTracer
      * @param ignorePassableBlocks If the raytrace should ignore blocks that are passable (grass, flowers, etc)
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Location rayTraceLocation(Location location, int maxDistance, FluidCollisionMode fluidCollisionMode, boolean ignorePassableBlocks)
+    public static Location rayTraceLocation(Location location, double maxDistance, FluidCollisionMode fluidCollisionMode, boolean ignorePassableBlocks)
     {
         World world = location.getWorld();
         Vector lookVec = location.getDirection();
@@ -385,7 +385,7 @@ public final class RayTracer
      * @param filter      The entity filter that should be used when raytracing
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Vector rayTraceVector(Location location, int maxDistance, Predicate<Entity> filter)
+    public static Vector rayTraceVector(Location location, double maxDistance, Predicate<Entity> filter)
     {
         return rayTraceVector(location, maxDistance, FluidCollisionMode.NEVER, false, filter);
     }
@@ -402,7 +402,7 @@ public final class RayTracer
      * @param filter             The entity filter that should be used when raytracing
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Vector rayTraceVector(Location location, int maxDistance, FluidCollisionMode fluidCollisionMode, Predicate<Entity> filter)
+    public static Vector rayTraceVector(Location location, double maxDistance, FluidCollisionMode fluidCollisionMode, Predicate<Entity> filter)
     {
         return rayTraceVector(location, maxDistance, fluidCollisionMode, false, filter);
     }
@@ -419,7 +419,7 @@ public final class RayTracer
      * @param filter               The entity filter that should be used when raytracing
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Vector rayTraceVector(Location location, int maxDistance, boolean ignorePassableBlocks, Predicate<Entity> filter)
+    public static Vector rayTraceVector(Location location, double maxDistance, boolean ignorePassableBlocks, Predicate<Entity> filter)
     {
         return rayTraceVector(location, maxDistance, FluidCollisionMode.NEVER, ignorePassableBlocks, filter);
     }
@@ -435,7 +435,7 @@ public final class RayTracer
      * @param filter The entity filter that should be used when raytracing
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Vector rayTraceVector(Location location, int maxDistance, FluidCollisionMode fluidCollisionMode, boolean ignorePassableBlocks, Predicate<Entity> filter)
+    public static Vector rayTraceVector(Location location, double maxDistance, FluidCollisionMode fluidCollisionMode, boolean ignorePassableBlocks, Predicate<Entity> filter)
     {
         World world = location.getWorld();
         Vector lookVec = location.getDirection();
@@ -473,7 +473,7 @@ public final class RayTracer
      * @param filter      The entity filter that should be used when raytracing
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Location rayTraceLocation(Location location, int maxDistance, Predicate<Entity> filter)
+    public static Location rayTraceLocation(Location location, double maxDistance, Predicate<Entity> filter)
     {
         return rayTraceLocation(location, maxDistance, FluidCollisionMode.NEVER, false, filter);
     }
@@ -490,7 +490,7 @@ public final class RayTracer
      * @param filter             The entity filter that should be used when raytracing
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Location rayTraceLocation(Location location, int maxDistance, FluidCollisionMode fluidCollisionMode, Predicate<Entity> filter)
+    public static Location rayTraceLocation(Location location, double maxDistance, FluidCollisionMode fluidCollisionMode, Predicate<Entity> filter)
     {
         return rayTraceLocation(location, maxDistance, fluidCollisionMode, false, filter);
     }
@@ -507,7 +507,7 @@ public final class RayTracer
      * @param filter               The entity filter that should be used when raytracing
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Location rayTraceLocation(Location location, int maxDistance, boolean ignorePassableBlocks, Predicate<Entity> filter)
+    public static Location rayTraceLocation(Location location, double maxDistance, boolean ignorePassableBlocks, Predicate<Entity> filter)
     {
         return rayTraceLocation(location, maxDistance, FluidCollisionMode.NEVER, ignorePassableBlocks, filter);
     }
@@ -523,7 +523,7 @@ public final class RayTracer
      * @param filter               The entity filter that should be used when raytracing
      * @return The Location of the raytrace hit, null if the raytrace never hit anything
      */
-    public static Location rayTraceLocation(Location location, int maxDistance, FluidCollisionMode fluidCollisionMode, boolean ignorePassableBlocks, Predicate<Entity> filter)
+    public static Location rayTraceLocation(Location location, double maxDistance, FluidCollisionMode fluidCollisionMode, boolean ignorePassableBlocks, Predicate<Entity> filter)
     {
         World world = location.getWorld();
         Vector lookVec = location.getDirection();
